feat(arrivals-slideout): move arrival rows loop to separate component
