+++ src/view/arrivals/components/ArrivalsTable.tsx
@@ -1,12 +1,11 @@
 import classNames from "classnames";
-import { map, sortBy } from "lodash";
 import { Moment } from "moment";
 import React from "react";
 import { Arrival, Route, StopLocation } from "../../../api/trimet/types";
-import ArrivalRow from "./ArrivalRow";
+import ArrivalRows from "./ArrivalRows";
 import "./Arrivals.css";
 
-export interface Props {
+interface Props {
   arrivals: Arrival[];
   loading: boolean;
   now: Moment;
@@ -14,47 +13,7 @@ export interface Props {
   stopLocation: StopLocation;
 }
 
-function sortArrivalsByEstimatedTime(arrivals: Arrival[]): Arrival[] {
-  return sortBy(arrivals, (arrival: Arrival) => arrival.estimated);
-}
-
-function getArrivalRoute(routes: Route[], routeId: number) {
-  return routes.find((route: Route) => route.route === routeId);
-}
-
 export default class ArrivalsTable extends React.Component<Props> {
-  public static getRows(
-    arrivals: Arrival[],
-    now,
-    onClick,
-    stopLocation: StopLocation
-  ) {
-    return map(sortArrivalsByEstimatedTime(arrivals), (arrival: Arrival) => {
-      const {
-        scheduled,
-        estimated,
-        feet,
-        route: routeId,
-        shortSign,
-        id
-      } = arrival;
-
-      return (
-        <ArrivalRow
-          key={id}
-          estimated={estimated}
-          feet={feet}
-          scheduled={scheduled}
-          routeId={routeId}
-          shortSign={shortSign}
-          now={now}
-          route={getArrivalRoute(stopLocation.route, routeId)}
-          onRouteIndicatorClick={onClick}
-        />
-      );
-    });
-  }
-
   public render() {
     const {
       arrivals,
@@ -84,9 +43,12 @@ export default class ArrivalsTable extends React.Component<Props> {
             <th>Distance</th>
           </tr>
         </thead>
-        <tbody>
-          {ArrivalsTable.getRows(arrivals, now, onClick, stopLocation)}
-        </tbody>
+        <ArrivalRows
+          arrivals={arrivals}
+          onClick={onClick}
+          stopLocation={stopLocation}
+          now={now}
+        />
       </table>
     );
   }
